Added in the report generator to the testing UI, tested and confirmed save and load from EDS works, tested and confirmed Conduct EDS files works, repairs for some workflow bugs are included as well.

+++ Testing Winform/Testing Winform/Forms/Main.vb
@@ -20,13 +20,15 @@ Imports DevExpress.Utils.Svg
 Imports DevExpress.Utils.Drawing
 Imports System.Xml
 Imports System.Xml.Serialization
+Imports SAPIReportGenerator
+Imports SAPI_Report_Generator_Editor
 
 Namespace UnitTesting
 
     Partial Public Class frmMain
         Public strcLocal As EDSStructure
         Public strcEDS As EDSStructure
-        Public testingVersion As String = "1.0.0.6"
+        Public testingVersion As String = "1.0.0.7"
 
 #Region "Object Declarations"
         'Public myUnitBases As New DataTransfererUnitBase
@@ -558,7 +560,7 @@ Namespace UnitTesting
                     btnProcess9.Click, btnProcess10.Click, btnProcess11.Click, btnProcess12.Click,
                     btnProcess13.Click, btnProcess14.Click, btnProcess15.Click, btnProcess16.Click,
                     btnProcess17.Click, btnProcess18.Click, btnProcess19.Click, btnProcess20.Click,
-                    btnProcess21.Click, btnProcess22.Click
+                    btnProcess21.Click, btnProcess22.Click, btnProcess23.Click
             If isopening Then Exit Sub
 
             ButtonclickToggle(Me.Cursor, Cursors.WaitCursor)
@@ -666,7 +668,11 @@ Namespace UnitTesting
                         (eriCount > 0 And tags(0).ToLower = "step3b") Or
                         (maeCount > 0 And tags(0).ToLower = "step3c") Or
                         (manCount > 0 And tags(0).ToLower = "step3d") Then
-                        answer = MsgBox(msg, vbCritical + vbYesNo, "Archive Files?")
+                        If forceAcrchiving Then
+                            answer = vbYes
+                        Else
+                            answer = MsgBox(msg, vbCritical + vbYesNo, "Archive Files?")
+                        End If
                     End If
 
                     If answer = vbNo Then
@@ -781,12 +787,12 @@ Namespace UnitTesting
                         End If
                     Next
 
-                Case "step6", "step10"
-                    Dim conductPath As String
+                Case "step6", "step11"
+                    Dim conductPath As String = ""
                     Select Case tags(0).ToLower
                         Case "step6"
                             conductPath = MaeFolder
-                        Case "step10"
+                        Case "step11"
                             conductPath = EDSFolder
                     End Select
 
@@ -841,9 +847,42 @@ Namespace UnitTesting
 
                     LogActivity("INFO | Results for all files compared and created in testing directory.")
                 Case "step8"
+                    'report generator
+                    Dim reportTemplate As String = "\\netapp4\common\Installers (Engineering Development)\SA Report Generator\Reference\Template.docx"
+                    Dim reportMapping As String = "\\netapp4\common\Installers (Engineering Development)\SA Report Generator\Reference\mapping.xml"
+
+                    Dim mylocation As String = DetermineFolder("Stop Report Generation")
+                    If mylocation = "STOP" Then
+                        LogActivity("INFO | Report generation cancelled")
+                        Exit Select
+                    End If
+
+                    CreateStructure(mylocation, False)
+                    SetStructureToPropertyGrid(strcLocal, pgcUnitTesting)
+
+                    Dim testReport As New Report(reportTemplate, reportMapping, strcLocal, My.Settings.booReportOption)
+
+                    Try
+                        Using reportEditor As New ReportEditorForm(testReport)
+                            reportEditor.ShowDialog()
+                        End Using
+                    Catch ex As Exception
+                        MsgBox(ex.Message, vbExclamation, "Failed to load report generator")
+                    End Try
+
+                Case "step9"
+                    Dim mylocation As String = DetermineFolder("Stop EDS Saving")
+                    If mylocation = "STOP" Then
+                        LogActivity("INFO | EDS Saving cancelled")
+                        Exit Select
+                    End If
+
+                    CreateStructure(mylocation, False)
+
                     strcLocal.SavetoEDS(EDSnewId, EDSdbActive)
                     LogActivity("INFO | EDS results attempted to save.")
-                Case "step9"
+                    SetStructureToPropertyGrid(strcLocal, pgcUnitTesting)
+                Case "step10"
                     Dim workingdirectory As String
 
                     'Check if eds folder exists
@@ -860,8 +899,8 @@ Namespace UnitTesting
                         DoArchiving(EDSFolder)
                     End If
 
-                    strcLocal.Clear()
-                    strcLocal = New EDSStructure(txtFndBU.Text, txtFndStrc.Text, txtFndWO.Text, EDSFolder, EDSFolder, EDSnewId, EDSdbActive)
+                    'strcLocal.Clear()
+                    strcLocal = New EDSStructure(testBu.Text, testSid.Text, testWo.Text, EDSFolder, EDSFolder, EDSnewId, EDSdbActive)
                     strcLocal.SaveTools(EDSFolder)
                     LogActivity("INFO | All files have been created in the directory '\Iteration " & iteration & "\EDS'.")
             End Select
@@ -870,6 +909,37 @@ finishMe:
             LogActivity("PROCESS | End " & sender.tooltip.ToString, True)
             ButtonclickToggle(Me.Cursor, Cursors.Default)
         End Sub
+
+        Private Function DetermineFolder(ByVal stopping As String) As String
+            Dim edsExists As Boolean = Directory.Exists(EDSFolder)
+            Dim whichFolder As New DialogResult
+            Dim maeOption As String = "YES = '\Maestro' Folder" & vbCrLf & vbCrLf
+            Dim manOption As String = "NO = '\Manual (SAPI)' Folder" & vbCrLf & vbCrLf
+            Dim cancelOption As String = "CANCEL = " & stopping & vbCrLf
+            Dim edsOption As String = "NO = '\EDS' Folder" & vbCrLf & vbCrLf
+            Dim filesPath As String
+
+            whichFolder = MsgBox("Which folder would you like use to create a report?" & vbCrLf & vbCrLf & IIf(edsExists, maeOption + edsOption + cancelOption, maeOption + manOption + cancelOption), vbYesNoCancel + vbInformation, "Which Folder?")
+
+            If whichFolder = vbCancel Then Return "STOP"
+
+            Select Case whichFolder
+                Case vbYes
+                    filesPath = MaeFolder
+                Case vbNo
+                    If edsExists Then
+                        filesPath = EDSFolder
+                        LogActivity("INFO | " & stopping.Replace("Stop ", "") & ": " & filesPath)
+                    Else
+                        filesPath = ManFolder
+                        LogActivity("INFO | " & stopping & ": " & filesPath)
+                    End If
+            End Select
+
+
+
+            Return filesPath
+        End Function
 
         Private Sub testBugFile_Click(sender As Object, e As EventArgs) Handles testBugFile.Click
             If Not Directory.Exists(Me.itFolder & "\Bug Reference Files") Then
@@ -1041,6 +1111,10 @@ StopLookingAtMeSwan:
             My.Settings.booConductDevMode = CheckEditDevMode.Checked
             My.Settings.Save()
         End Sub
+        Private Sub CheckEditAutoReport_CheckedChanged(sender As Object, e As EventArgs) Handles CheckEditAutoReport.CheckedChanged
+            My.Settings.booReportOption = CheckEditAutoReport.Checked
+            My.Settings.Save()
+        End Sub
 
         Private Sub CheckEditExcelVisible_CheckedChanged(sender As Object, e As EventArgs) Handles CheckEditExcelVisible.CheckedChanged
             My.Settings.booConductExcelVis = CheckEditExcelVisible.Checked
@@ -1052,6 +1126,47 @@ StopLookingAtMeSwan:
             My.Settings.Save()
         End Sub
 
+        Public forceAcrchiving As Boolean = False
+        Private Sub btnAuto_Click(sender As Object, e As EventArgs) Handles btnAuto.Click
+            ButtonclickToggle(Me.Cursor)
+            forceAcrchiving = True
+            For i As Integer = 0 To 99
+                'Test cases we removed and guyed towers with LRT
+                If Not (i = 87 Or i = 76 Or i = 52 Or i = 51 Or i = 49 Or i = 91 Or i = 92) Then
+                    'Set up iteration folder
+                    testID.SelectedIndex = i
+                    'Create Iteration folder
+                    If Not Directory.Exists(Me.dirUse & "\Test ID " & Me.testCase & "\Iteration " & Me.iteration) Then
+                        CreateIteration(Me.iteration)
+                    End If
+                    'Create MAE files
+                    TestSteps(btnProcess11, EventArgs.Empty)
+                    'Create Man files just in case compare results needs it
+                    TestSteps(btnProcess13, EventArgs.Empty)
+                    'Import inputs for MAE files
+                    TestSteps(btnProcess5, EventArgs.Empty)
+                    'Conduct MAE files
+                    TestSteps(btnProcess7, EventArgs.Empty)
+                    'Create excel results (MAE v. CUR only, but everything will be created)
+                    TestSteps(btnProcess21, EventArgs.Empty)
+                End If
+            Next
+            forceAcrchiving = False
+            ButtonclickToggle(Me.Cursor)
+        End Sub
+
+        Private Sub btnChecking_Click(sender As Object, e As EventArgs) Handles btnChecking.Click
+            ButtonclickToggle(Me.Cursor)
+            For i As Integer = 0 To 99
+                'Test cases we removed and guyed towers with LRT
+                If Not (i = 87 Or i = 76 Or i = 52 Or i = 51 Or i = 49 Or i = 91 Or i = 92) Then
+                    'Pull all files and check it out if possible.
+                    PullIt(True, True)
+                End If
+            Next
+            SetTestIDLabels()
+            ButtonclickToggle(Me.Cursor)
+        End Sub
 #End Region
 
 #Region "My Largely Little Helpers"
@@ -1150,11 +1265,13 @@ StopLookingAtMeSwan:
 #End Region
 
 #Region "Checkin/Checkout"
-        Public Sub PullIt(ByVal checkout As Boolean)
+        Public Sub PullIt(ByVal checkout As Boolean, Optional ByVal checkingOutEverything As Boolean = False)
             ForceSyncing(SyncDirection.RtoLocal, True)
 
             If Not testPush.Enabled Then
-                File.Delete(Me.dirUse & "\Test ID " & Me.testCase & "\Checked Out.txt")
+                If Not checkingOutEverything Then
+                    File.Delete(Me.dirUse & "\Test ID " & Me.testCase & "\Checked Out.txt")
+                End If
             End If
 
             If checkout Then
@@ -1340,6 +1457,7 @@ StopLookingAtMeSwan:
             btnProcess20.Enabled = Not btnProcess20.Enabled
             btnProcess21.Enabled = Not btnProcess21.Enabled
             btnProcess22.Enabled = Not btnProcess22.Enabled
+            btnProcess23.Enabled = Not btnProcess23.Enabled
 
             XtraTabControl1.Enabled = Not XtraTabControl1.Enabled
             testBugFile.Enabled = Not testBugFile.Enabled
@@ -1350,6 +1468,7 @@ StopLookingAtMeSwan:
             CheckEditDevMode.Enabled = Not CheckEditDevMode.Enabled
             CheckEditExcelVisible.Enabled = Not CheckEditExcelVisible.Enabled
             CheckEditExcelVisibleII.Enabled = Not CheckEditExcelVisibleII.Enabled
+            CheckEditAutoReport.Enabled = Not CheckEditAutoReport.Enabled
         End Sub
 #End Region
 
@@ -2209,7 +2328,7 @@ StopLookingAtMeSwan:
         End Sub
 
         'Creates a structure object based on the files in the maestro folder for the current iteration
-        Public Sub CreateStructure(ByVal conductPath As String)
+        Public Sub CreateStructure(ByVal filesPath As String, Optional ByVal deleteAdditionalTNXfiles As Boolean = True)
             Dim myFiles As String()
             Dim myFilesLst As New List(Of String)
 
@@ -2217,7 +2336,7 @@ StopLookingAtMeSwan:
             'Dim response As DialogResult = DialogResult.Cancel
 
             'Loop through all files in the maestro folder for the current test case and iteration
-            For Each info As FileInfo In New DirectoryInfo(conductPath).GetFiles
+            For Each info As FileInfo In New DirectoryInfo(filesPath).GetFiles
                 If info.Extension = ".eri" Then
                     'All eris permitted
                     myFilesLst.Add(info.FullName)
@@ -2241,15 +2360,17 @@ StopLookingAtMeSwan:
                     '    End If
                     'End If
                     'If response = DialogResult.Yes Then
-                    info.Delete()
-                    LogActivity("DEBUG | File Deleted: " & info.FullName)
+                    If deleteAdditionalTNXfiles Then
+                        info.Delete()
+                        LogActivity("DEBUG | File Deleted: " & info.FullName)
+                    End If
                     'End If
                 End If
             Next
 
             'Convert the list of valid file names to an array for creating anew structure
             myFiles = myFilesLst.ToArray
-            strcLocal = New EDSStructure(testBu.Text, testSid.Text, testWo.Text, Me.MaeFolder, Me.MaeFolder, myFiles, EDSnewId, EDSdbActive)
+            strcLocal = New EDSStructure(testBu.Text, testSid.Text, testWo.Text, filesPath, filesPath, myFiles, EDSnewId, EDSdbActive)
         End Sub
 
         'Loads the CSV with the test cases 
@@ -2393,9 +2514,6 @@ StopLookingAtMeSwan:
 
             Return returner
         End Function
-
-
-
